requirements.txt
Flask==3.1.0
flask_caching==2.3.0
numpy==2.2.2
openai==1.59.8
Pillow==11.1.0
python-dotenv==1.0.1
pyocclient==0.6
opencv-python==4.11.0.86
